# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StationPlay API",
    "description": "StationPlay's API: the server, its stations, their guide and streams, and things an Admin can have it do. Send an API token (made on StationPlay's Access tab) as `Authorization: Bearer <token>`: a Viewer token reads, an Admin token also does what's marked Admin. Version 1 only grows; see docs/api.md.",
    "version": "1",
    "license": {
      "name": "GPL-3.0-only"
    }
  },
  "paths": {
...
    "/api/v1/status": {
      "get": {
        "summary": "How StationPlay is doing",
        "description": "Whether Plex answers, how many tuners there are, and what's playing.",
        "operationId": "getStatus",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations/{number}/update": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Update a station from Plex now",
        "description": "Follows Plex straight away (as Update now does on the page): new and removed programs, with a held update applied. Admin.",
        "operationId": "updateStation",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Updated"
                }
              }
            }
          },
          "400": {
            "description": "Something in the request wasn't understood",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "There's no such station",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations/{number}/check": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Check a station's files",
        "description": "Starts the quick check of each of its programs' files, as Check files does on the page (or, while one is running, says how far it has got). Admin.",
        "operationId": "checkStation",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Check"
                }
              }
            }
          },
          "400": {
            "description": "Something in the request wasn't understood",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "There's no such station",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "How a station's file check is going",
        "description": "Its last check, running or done, without starting one. Admin.",
        "operationId": "getStationCheck",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Check"
                }
              }
            }
          },
          "400": {
            "description": "Something in the request wasn't understood",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "There's no such station",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guide/refresh": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Ask Plex to refresh its guide",
        "description": "As StationPlay does by itself after a change. Admin.",
        "operationId": "refreshGuide",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideRefreshed"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backups": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Back StationPlay up now",
        "description": "Makes a backup, as StationPlay does every night. Admin.",
        "operationId": "makeBackup",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Backup"
                }
              }
            }
          },
          "401": {
            "description": "No token, or one that isn't valid (while signing in is on)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token may not do this, or isn't accepted from there",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Backup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The backup's file name, in the backups folder"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Backup"
      },
      "Check": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "running": {
            "type": "boolean",
            "title": "Running",
            "description": "Whether the check is still going"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Programs to check (0: no check yet)"
          },
          "done": {
            "type": "integer",
            "title": "Done",
            "description": "Programs checked so far"
          },
          "newlyBroken": {
            "type": "integer",
            "title": "Newlybroken",
            "description": "Programs it found problems in (now off the air)"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "description": "Programs that couldn't be checked"
          }
        },
        "type": "object",
        "required": [
          "number",
          "running",
          "total",
          "done",
          "newlyBroken",
          "skipped"
        ],
        "title": "Check"
      },
...
      "GuideRefreshed": {
        "properties": {
          "refreshed": {
            "type": "boolean",
            "title": "Refreshed",
            "description": "Whether Plex said it would refresh its guide"
          }
        },
        "type": "object",
        "required": [
          "refreshed"
        ],
        "title": "GuideRefreshed"
      },
...
      "Playing": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number",
            "description": "The station's number"
          },
          "viewers": {
            "type": "integer",
            "title": "Viewers",
            "description": "Devices watching it"
          }
        },
        "type": "object",
        "required": [
          "number",
          "viewers"
        ],
        "title": "Playing"
      },
...
      "Status": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version",
            "description": "StationPlay's version"
          },
          "plexConnected": {
            "type": "boolean",
            "title": "Plexconnected",
            "description": "Whether Plex answers"
          },
          "tuners": {
            "type": "integer",
            "title": "Tuners",
            "description": "How many stations can play at once"
          },
          "playing": {
            "items": {
              "$ref": "#/components/schemas/Playing"
            },
            "type": "array",
            "title": "Playing",
            "description": "The stations playing now"
          }
        },
        "type": "object",
        "required": [
          "version",
          "plexConnected",
          "tuners",
          "playing"
        ],
        "title": "Status"
      },
      "Updated": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "changed": {
            "type": "boolean",
            "title": "Changed",
            "description": "Whether Plex had anything new for it"
          }
        },
        "type": "object",
        "required": [
          "number",
          "changed"
        ],
        "title": "Updated"
      },
      "Error": {
        "type": "object",
        "title": "Error",
        "properties": {
          "detail": {
            "type": "string",
            "description": "Why, as a sentence to show"
          }
        },
        "required": [
          "detail"
        ]
      }
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token from the Access tab (or a StationPlay app's sign-in)"
      }
    }
  },
...
  "security": [
    {
      "token": []
    }
  ]
}